Green tests for instance sagas publishing fault on transition error Namespaces adjustments Move classes to own files
