+++ GridDomain.Domain.Tests/Aggregate_Sagas_actor_lifetime/Infrastructure/InstanceSagaPersistedHub_Infrastructure.cs
@@ -4,7 +4,6 @@ using Akka.DI.Core;
 using GridDomain.EventSourcing.Sagas;
 using GridDomain.EventSourcing.Sagas.InstanceSagas;
 using GridDomain.Node.Actors;
-using GridDomain.Tests.Aggregate_Sagas_actor_lifetime.Actors;
 using GridDomain.Tests.Sagas.InstanceSagas;
 using GridDomain.Tests.Sagas.SoftwareProgrammingDomain.Events;
