+++ GridDomain.Domain.Tests/Sagas/InstanceSagas/Given_saga_When_handling_command_faults.cs
@@ -7,6 +7,7 @@ using GridDomain.CQRS.Messaging;
 using GridDomain.EventSourcing.Sagas;
 using GridDomain.EventSourcing.Sagas.InstanceSagas;
 using GridDomain.Node.Configuration.Composition;
+using GridDomain.Tests.Sagas.SoftwareProgrammingDomain;
 using GridDomain.Tests.Sagas.SoftwareProgrammingDomain.Commands;
 using GridDomain.Tests.Sagas.SoftwareProgrammingDomain.Events;
 using NUnit.Framework;
