+++ GridDomain.Domain.Tests/Sagas/SoftwareProgrammingDomain/CantFindSofaException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace GridDomain.Tests.Sagas.SoftwareProgrammingDomain.Commands
+namespace GridDomain.Tests.Sagas.SoftwareProgrammingDomain
 {
     public class CantFindSofaException : Exception
     {
