+++ GridDomain.Domain.Tests/Sagas/StateSagas/SampleSaga/UndefinedCoffeMachineException.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace GridDomain.Tests.Sagas.StateSagas.SampleSaga
+{
+    public class UndefinedCoffeMachineException : Exception
+    {
+    }
+}
