+++ GridDomain.Domain.Tests/Sagas/StateSagas/When_state_saga_raises_an_exception.cs
@@ -2,8 +2,6 @@ using System;
 using GridDomain.CQRS;
 using GridDomain.CQRS.Messaging;
 using GridDomain.EventSourcing.Sagas;
-using GridDomain.EventSourcing.Sagas.InstanceSagas;
-using GridDomain.Tests.Sagas.InstanceSagas;
 using GridDomain.Tests.Sagas.SoftwareProgrammingDomain.Events;
 using GridDomain.Tests.Sagas.StateSagas.SampleSaga;
 using Microsoft.Practices.Unity;
@@ -15,7 +13,7 @@ namespace GridDomain.Tests.Sagas.StateSagas
     [TestFixture]
     public class When_state_saga_raises_an_exception : SoftwareProgramming_StateSaga_Test
     {
-        protected override TimeSpan Timeout { get; } = TimeSpan.FromSeconds(1000);
+        protected override TimeSpan Timeout { get; } = TimeSpan.FromSeconds(3);
 
         [SetUp]
         public void When_saga_receives_a_message_that_case_saga_exception()
@@ -24,16 +22,11 @@ namespace GridDomain.Tests.Sagas.StateSagas
             var personId = Guid.NewGuid();
 
             //prepare initial saga state
-            var sagaData = new SoftwareProgrammingSagaData(nameof(InstanceSagas.SoftwareProgrammingSaga.MakingCoffee))
-            {
-                PersonId = personId
-            };
-            var sagaDataEvent = new SagaCreatedEvent<SoftwareProgrammingSagaData>(sagaData, sagaId);
-            SaveInJournal<SagaDataAggregate<SoftwareProgrammingSagaData>>(sagaId, sagaDataEvent);
+            
+            var sagaDataEvent = new SagaCreatedEvent<SoftwareProgrammingSaga.States>(SoftwareProgrammingSaga.States.MakingCoffe, sagaId);
+            SaveInJournal<SoftwareProgrammingSagaState>(sagaId, sagaDataEvent);
 
-
-            var publisher = GridNode.Container.Resolve<IPublisher>();
-            publisher.Publish(new CoffeMakeFailedEvent(Guid.Empty,personId).CloneWithSaga(sagaId));
+            Publisher.Publish(new CoffeMakeFailedEvent(Guid.Empty,personId).CloneWithSaga(sagaId));
         }
 
         [Test]
@@ -43,16 +36,23 @@ namespace GridDomain.Tests.Sagas.StateSagas
         }
 
         [Test]
+        public void Fault_exception_should_contains_stack_trace()
+        {
+            var fault = (IFault<CoffeMakeFailedEvent>)WaitFor<IFault<CoffeMakeFailedEvent>>(false).Message;
+            Assert.True(fault.Exception.UnwrapInner().StackTrace.Contains(typeof(SoftwareProgrammingSaga).Name));
+        }
+
+        [Test]
         public void Fault_should_have_saga_as_producer()
         {
             var fault = (IFault<CoffeMakeFailedEvent>)WaitFor<IFault<CoffeMakeFailedEvent>>(false).Message;
-            Assert.AreEqual(fault.Processor, typeof(SoftwareProgrammingSaga));
+            Assert.AreEqual(typeof(SoftwareProgrammingSaga), fault.Processor);
         }
         [Test]
         public void Fault_should_contains_exception_from_saga()
         {
             var fault = (IFault<CoffeMakeFailedEvent>)WaitFor<IFault<CoffeMakeFailedEvent>>(false).Message;
-            Assert.IsInstanceOf<UndefinedCoffeMachineException>(fault.Exception);
+            Assert.IsInstanceOf<UndefinedCoffeMachineException>(fault.Exception.UnwrapInner());
         }
     }
 }
